Feat: Create an internal role to generate ansible.cfg variants
This role can be used to generate an ansible.cfg config file from a
template. It would be useful to generate various flavors of this config
file, especially when you want to run playbooks from Vagrant or from
Jenkins, as the settings may differ depending the system.

At this stage, this role only copy the original ansible.cfg file, and
make some modifications in it, especially removes SSH section and update
paths.